Features added -Removed GSON converter -Used Kotlin Serialization -Manual dependency items simplified

+++ Mani-Quotify/app/src/main/java/com/mani/quotify007/domain/model/QuoteResult.kt
@@ -1,10 +1,10 @@
 package com.mani.quotify007.domain.model
 
 data class QuoteResult(
-    val count: Int,
-    val lastItemIndex: Int,
-    val page: Int,
+    val count: Long,
+    val lastItemIndex: Long,
+    val page: Long,
     val results: List<Quote>,
-    val totalCount: Int,
-    val totalPages: Int
+    val totalCount: Long,
+    val totalPages: Long
 )
